requirements.txt
fastapi
uvicorn[standard]

pandas 
numpy
colorlog
toml

sqlparse
sqlglot
peewee

pyodbc
mysqlclient
cx_oracle
pymongo
clickhouse-connect
clickhouse-sqlalchemy